Reorganize the page files as per review comment
