+++ sonic_platform_base/sonic_xcvr/mem_maps/public/cmis/pages/__init__.py
@@ -1,11 +1,11 @@
 """
-    cmis_pages package
+    cmis.pages package
 
     CMIS page-specific memory map classes
 """
 
-from .base import CmisPage
-from .cmis_page_consts import (
+from .page import CmisPage
+from .consts import (
     CMIS_NUM_NON_BANKED_PAGES,
     CMIS_EEPROM_PAGE_SIZE,
     ADMINISTRATIVE_PAGE,
@@ -18,16 +18,16 @@ from .cmis_page_consts import (
     VDM_ADVERTISING_CTRL_PAGE,
     CDB_MESSAGE_PAGE,
 )
-from .pg_00_administrative_lower import CmisAdministrativeLowerPage
-from .pg_00_administrative_upper import CmisAdministrativeUpperPage
-from .pg_01_advertising import CmisAdvertisingPage
-from .pg_02_thresholds import CmisThresholdsPage
-from .pg_10_lane_datapath_config import CmisLaneDatapathConfigPage
-from .pg_11_lane_datapath_status import CmisLaneDatapathStatusPage
-from .pg_12_tunable_laser_ctrl_status import CmisTunableLaserCtrlStatusPage
-from .pg_13_module_perf_diag_ctrl import CmisModulePerfDiagCtrlPage
-from .pg_2f_vdm_advertising_ctrl import CmisVdmAdvertisingCtrlPage
-from .pg_9f_cdb_message import CmisCdbMessagePage
+from .page00_lower import CmisAdministrativeLowerPage
+from .page00_upper import CmisAdministrativeUpperPage
+from .page01 import CmisAdvertisingPage
+from .page02 import CmisThresholdsPage
+from .page10 import CmisLaneDatapathConfigPage
+from .page11 import CmisLaneDatapathStatusPage
+from .page12 import CmisTunableLaserCtrlStatusPage
+from .page13 import CmisModulePerfDiagCtrlPage
+from .page2f import CmisVdmAdvertisingCtrlPage
+from .page9f import CmisCdbMessagePage
 
 __all__ = [
     'CmisPage',
