+++ sonic_platform_base/sonic_xcvr/mem_maps/public/cmis/pages/consts.py
